address comments

+++ pkg/vm/memory/memory_value.go
@@ -342,10 +342,10 @@ func (mv *MemoryValue) addrUnsafe() *MemoryAddress {
 	return (*MemoryAddress)(unsafe.Pointer(&mv.Felt))
 }
 
-func (memory *Memory) GetConsecutiveMemoryValues(addr MemoryAddress, size int16) ([]MemoryValue, error) {
+func (memory *Memory) GetConsecutiveMemoryValues(addr MemoryAddress, size uint64) ([]MemoryValue, error) {
 	values := make([]MemoryValue, size)
-	for i := int16(0); i < size; i++ {
-		nAddr, err := addr.AddOffset(i)
+	for i := uint64(0); i < size; i++ {
+		nAddr, err := addr.AddOffset(int16(i))
 		if err != nil {
 			return nil, err
 		}
@@ -359,7 +359,7 @@ func (memory *Memory) GetConsecutiveMemoryValues(addr MemoryAddress, size int16)
 }
 
 func (memory *Memory) ResolveAsBigInt3(valAddr MemoryAddress) ([3]*f.Element, error) {
-	valMemoryValues, err := memory.GetConsecutiveMemoryValues(valAddr, int16(3))
+	valMemoryValues, err := memory.GetConsecutiveMemoryValues(valAddr, uint64(3))
 	if err != nil {
 		return [3]*f.Element{}, err
 	}
@@ -377,7 +377,7 @@ func (memory *Memory) ResolveAsBigInt3(valAddr MemoryAddress) ([3]*f.Element, er
 }
 
 func (memory *Memory) ResolveAsEcPoint(valAddr MemoryAddress) ([2]*f.Element, error) {
-	valMemoryValues, err := memory.GetConsecutiveMemoryValues(valAddr, int16(2))
+	valMemoryValues, err := memory.GetConsecutiveMemoryValues(valAddr, uint64(2))
 	if err != nil {
 		return [2]*f.Element{}, err
 	}
